ENG-3922 J1: stop charging writePrice for the test's own struct build
_reportWrite evaluated _params INSIDE the measured span, so the struct build
and its two keccak256 provenance hashes were billed to writePrice, and it
called _params a SECOND time for the calldata, advancing nextCycle so the
payload costed was not the payload written. The params are now built once and
hoisted above vm.cool.

Four baseline write rows move; the heartbeat control does not, which is the
check that the fix is the fix:

  write 1, new row          149,310 -> 146,852   -2,458
  write 2, slot+counter     133,166 -> 130,697   -2,469
  writes 3-48               116,079 -> 113,600   -2,479
  write 49+, wrapped ring    82,509 ->  79,501   -3,008
  heartbeat (control)         23,040 -> 23,040        0

Adding the bounds guard CodeRabbit asked for on the EAS recipient truncation
made the harness fail its own test, and the reason is worth more than the
guard. FabricaToken derives ids as uint64(keccak256(...)) widened
(FabricaToken.sol:363-365), so every real id is below 2^64 with 96 bits of
headroom -- the truncation is lossless in production. The HARNESS was using
full 256-bit keccak ids, which is not what production looks like, and is
exactly the shape that would have hidden the truncation this arm depends on.
Ids are now Fabrica-shaped throughout.

That moves calldata, in the safe direction: EIP-2028 charges 4 gas for a zero
byte against 16 for a non-zero, and a uint64 id carries 24 zero bytes where a
256-bit id carries almost none. About 288 gas per id, so pointer pointBatch
per item 27,016 -> 26,728, ownerless writePriceBatch 72,794 -> 72,507, EAS
multiAttest 260,123 -> 259,692. It lands on every arm in proportion, so no
comparison moves, and the previously recorded Sepolia write costs are high by
that much rather than low. Recorded in the evidence file rather than papered
over.

Arm 1's totals now include the Indexer writes its own lookups require: an EAS
record it cannot find is a record it does not have. Coverage stamp with
indexing is 353,292 per item against the pointer arm's 25,283 -- fourteen
times -- and the cycle close is 469,623. That sharpens the round-3 finding
rather than changing it.

Also: the lock test now covers ArmEasIndexer, whose Indexer-based head lookup
is a separate candidate path the comment already claimed was covered; and
derive.py emits a blank line around generated tables (MD058).

+++ test/bench/Eng3922Sepolia.s.sol
@@ -128,7 +128,7 @@ contract Eng3922SepoliaScript is Script {
     }
 
     function _tokenIdAt(uint256 i) internal pure returns (uint256) {
-        return uint256(keccak256(abi.encode("eng3922-sepolia-token", i)));
+        return uint256(uint64(uint256(keccak256(abi.encode("eng3922-sepolia-token", i)))));
     }
 
     /// @notice The audit commitment over the cycle's token ids. Nothing on chain verifies it.
